requirements.txt
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
Flask==3.1.0
getstream==1.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==3.26.1
mypy-extensions==1.0.0
packaging==24.2
PyJWT==2.10.1
python-dotenv==1.0.1
sniffio==1.3.1
typing-inspect==0.9.0
typing_extensions==4.12.2
Werkzeug==3.1.3
